snapshot before rtree based inpoly test
This commit contains tests to perform the inpoly test in parallel using
path.contains_point(). Parallel processing of path.contains_point() does not
seem to privde much benefit over using path.contains_points() directly. Therefore, the parallel processing of this call will be eliminated, and an rtree based filtering will be implemented to sieve first the centroids of the clipped size function to the rtree index of the edges of the pslg, subsantially reducing the amount of points to pass to the inpoly test.